Fix/wpt rfid

+++ src/main/helpers/launch_wpt.js
@@ -10,10 +10,14 @@ module.exports = function launchWpt(wpt, callback) {
 
 	return new Promise((resolve, reject) => {
 		let timeout = setTimeout(() => {
-			child.stdout.removeAllListeners()
-			child.stderr.removeAllListeners()
+			if (child.stdout) {
+				child.stdout.removeAllListeners()
+			}
+			if (child.stderr) {
+				child.stderr.removeAllListeners()
+			}
 			child.removeAllListeners()
-			if (!child.killed) {
+			if (child && !child.killed) {
 				child.kill('SIGKILL')
 			}
 			if (wptPid) {
@@ -29,12 +33,10 @@ module.exports = function launchWpt(wpt, callback) {
 			)
 		}, 1000 * 20)
 		// cannot use fork same node version of nw used
-		const fork = require('child_process').fork
-		const execFile = require('child_process').execFile
+		const spawn = require('child_process').spawn
+		// const execFile = require('child_process').execFile
 
-		const options = {
-			stdio: ['pipe', 'pipe', 'pipe']
-		}
+
 
 		const isScript =
 			path.extname(wpt.path) === '.sh' || path.extname(wpt.path) === '.bat'
@@ -46,7 +48,7 @@ module.exports = function launchWpt(wpt, callback) {
 		const exe = isScript ? wpt.path : 'node'
 		const args = isScript
 			? []
-			: ['--experimental-worker', '--no-warnings', exePath]
+			: [exePath]
 
 		if (!fs.existsSync(exePath)) {
 			reject(
@@ -58,16 +60,26 @@ module.exports = function launchWpt(wpt, callback) {
 				)
 			)
 		}
-		if (isScript && path.extname(exePath) === '.sh' || isJs) {
-			// not working on Windows with .bat ...
-			options.stdio.push('ipc')
-		}
+
 
 		if (!isJs && path.extname(exePath) === '.bat') {
 			wpt.wait_on_ipc = false
 		}
 
-		const child = execFile(exe, args, options)
+		const options = {
+			stdio: wpt.wait_on_ipc ? ['pipe', 'pipe', 'pipe']: undefined,
+			windowsHide: true,
+			shell:wpt.shell,
+			detached: wpt.detached,
+		}
+
+		if (wpt.wait_on_ipc && options.stdio && isScript && (path.extname(exePath) === '.sh' || isJs)) {
+			// not working on Windows with .bat ...
+			options.stdio.push('ipc')
+		}
+
+		const child = spawn(exe, args, options)
+
 		if (wpt.wait_on_ipc) {
 			child.on('message', message => {
 				log.info('wpt.send', message)
@@ -84,8 +96,12 @@ module.exports = function launchWpt(wpt, callback) {
 						clearTimeout(timeout)
 						timeout = null
 					}
-					child.stdout.removeAllListeners()
-					child.stderr.removeAllListeners()
+					if (child.stdout) {
+						child.stdout.removeAllListeners()
+					}
+					if (child.stderr) {
+						child.stderr.removeAllListeners()
+					}
 					child.removeAllListeners()
 					resolve(child)
 				}
@@ -93,8 +109,8 @@ module.exports = function launchWpt(wpt, callback) {
 		}
 
 		if (
-			!wpt.wait_on_ipc ||
-			(process.env.DEBUG && process.env.DEBUG === 'wpt')
+			child.stdout && (!wpt.wait_on_ipc ||
+			(process.env.DEBUG && process.env.DEBUG === 'wpt'))
 		) {
 			child.stdout.on('data', function (data) {
 				if (process.env.DEBUG && process.env.DEBUG === 'wpt') {
@@ -104,6 +120,36 @@ module.exports = function launchWpt(wpt, callback) {
 				if (messages.length > 0) {
 					messages.length = ""
 				}
+
+				if (!wpt.wait_on_ipc && data.indexOf('[pid] ') >= 0) {
+					let pid = typeof data === "object" ? data.toString().split("\n") : data.split("\n")
+
+					for (let i = 0; i < pid.length; i++) {
+						if (pid[i].indexOf('[pid]' >= 0)) {
+							pid = pid[i]
+							break
+						}
+					}
+					const pids = pid.split(" ")
+					if (pid.length > 0) {
+						pid = pids.pop()
+						pid = Number.parseInt(pid, 10)
+						if (Number.isNaN(pid)) {
+							if (pids.length > 0) {
+								pid = pids.pop()
+								pid = Number.parseInt(pid, 10)
+							} else {
+								pid = null
+							}
+						}
+
+						if (pid && !Number.isNaN(pid) && callback) {
+							callback('get_wpt_pid_done', pid)
+						}
+					}
+
+				}
+
 				if (
 					!wpt.wait_on_ipc &&
 					(data.indexOf('[HTTP Server] started on port') >= 0 ||
@@ -120,38 +166,20 @@ module.exports = function launchWpt(wpt, callback) {
 				}
 			})
 		}
-
-		child.stderr.on('data', function (data) {
-			//   if (messages.length === 0) {
-			//     setTimeout(() => {
-			//       child.kill('SIGKILL')
-			//       if (wptPid) {
-			//         try {
-			//           process.kill(wptPid)
-			//         } catch (err2) {
-			//           // console.log(err2)
-			//         }
-			//       }
-			//       child.stdout.removeAllListeners()
-			//       child.stderr.removeAllListeners()
-			//       child.removeAllListeners()
-			//       const err = new CustomError(
-			//         400,
-			//         CustomError.CODE.WPT_CREATION_FAILED,
-			//         wpt.path,
-			//         []
-			//       )
-			//       err.messages = messages
-			//       reject(err)
-			//     }, 1000)
-			//   }
-			messages += data.toString()
-		})
-
+		if (child.stderr) {
+			child.stderr.on('data', function (data) {
+				messages += data.toString()
+			})
+		}
+		
 		child.once('exit', reason => {
 			setTimeout(() => {
-				child.stdout.removeAllListeners()
-				child.stderr.removeAllListeners()
+				if (child.stdout) {
+					child.stdout.removeAllListeners()
+				}
+				if (child.stderr) {
+					child.stderr.removeAllListeners()
+				}
 				child.removeAllListeners()
 				reject(
 					new CustomError(
@@ -172,12 +200,33 @@ module.exports = function launchWpt(wpt, callback) {
 			}
 			if (!child.killed) {
 				child.kill('SIGKILL')
-				child.stdout.removeAllListeners()
-				child.stderr.removeAllListeners()
+				if (child.stdout) {
+					child.stdout.removeAllListeners()
+				}
+				if (child.stderr) {
+					child.stderr.removeAllListeners()
+				}
 				child.removeAllListeners()
 			}
 			err.messages = messages
 			reject(err)
 		})
+
+		if (child && wpt.shell) {
+			if (timeout) {
+				clearTimeout(timeout)
+				timeout = null
+			}
+			setTimeout(() => {
+				if (child.stdout) {
+					child.stdout.removeAllListeners()
+				}
+				if (child.stderr) {
+					child.stderr.removeAllListeners()
+				}
+				child.removeAllListeners()
+				resolve(child)
+			}, 3000)
+		}
 	})
 }
